Separate TextShaderEditor and ShaderEditorPlugin into different files
